Move handler & merge file
